# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Casenova API Documentation",
        "description": "Casenova API Documentation",
        "contact": {
            "email": "admin@casenova.uz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://casenova.lc/api/v1",
            "description": "Api Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User authorization",
                "description": "Return User token",
                "operationId": "authUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/blog": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get list of blogs",
                "description": "Returns list of blogs",
                "operationId": "getBlogs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BlogCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Create new Blog",
                "description": "Returns Blog data",
                "operationId": "storeBlog",
                "requestBody": {
                    "description": "Pass blog credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBlogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/blog/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get blog information",
                "description": "Returns blog data",
                "operationId": "getBlogById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BlogResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update blog",
                "description": "Returns blog data",
                "operationId": "updateBlog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBlogRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/blog/{blog}/image": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update image for blog",
                "description": "Returns blog status",
                "operationId": "imageBlog",
                "parameters": [
                    {
                        "name": "blog",
                        "in": "path",
                        "description": "Blog id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BaseModel/properties/property_id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "$ref": "#/components/schemas/BaseModel/properties/property_uploads"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseModel/properties/property_boolean_result"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreBlogRequest": {
                "title": "StoreBlogRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/StoreBlogTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/StoreBlogTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/StoreBlogTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreBlogRequest"
                }
            },
            "StoreBlogTranslationRequest": {
                "title": "StoreBlogTranslationRequest",
                "properties": {
                    "title": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/title"
                    },
                    "content": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/content"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreBlogTranslationRequest"
                }
            },
...
            "UpdateBlogRequest": {
                "title": "UpdateBlogRequest",
                "properties": {
                    "ru": {
                        "$ref": "#/components/schemas/UpdateBlogTranslationRequest"
                    },
                    "uz": {
                        "$ref": "#/components/schemas/UpdateBlogTranslationRequest"
                    },
                    "en": {
                        "$ref": "#/components/schemas/UpdateBlogTranslationRequest"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateBlogRequest"
                }
            },
            "UpdateBlogTranslationRequest": {
                "title": "UpdateBlogTranslationRequest",
                "properties": {
                    "title": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/title"
                    },
                    "content": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/content"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateBlogTranslationRequest"
                }
            },
...
            "BlogCollection": {
                "title": "BlogCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BlogResource"
                        }
                    }
                },
                "type": "object"
            },
            "BlogResource": {
                "title": "BlogResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/Blog/properties/id"
                    },
                    "image": {
                        "$ref": "#/components/schemas/Blog/properties/image"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/Blog/properties/created_at"
                    },
                    "updatedAt": {
                        "$ref": "#/components/schemas/Blog/properties/updated_at"
                    },
                    "translations": {
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/BlogTranslationResource"
                                }
                            ]
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BlogResource"
                }
            },
...
            "BlogTranslationResource": {
                "title": "BlogTranslationResource",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/id"
                    },
                    "blogId": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/blog_id"
                    },
                    "locale": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/locale"
                    },
                    "title": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/title"
                    },
                    "content": {
                        "$ref": "#/components/schemas/BlogTranslation/properties/content"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BlogTranslationResource"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}